Fix single `_default` sitemap.
